Cleaned up useless state and got SHC images working again
