+++ aliv.cbp
@@ -10,7 +10,7 @@
 				<Option output="build/aliv" prefix_auto="1" extension_auto="1" />
 				<Option working_dir="build" />
 				<Option object_output="output/" />
-				<Option type="1" />
+				<Option type="0" />
 				<Option compiler="gcc" />
 				<Compiler>
 					<Add option="-Wall" />
@@ -28,36 +28,38 @@
 			<Add option="-Wall" />
 		</Compiler>
 		<Unit filename="include/BmpLoader.h" />
-		<Unit filename="include/GM1Loader.h" />
+		<Unit filename="include/GM1Parser.h" />
 		<Unit filename="include/GifLoader.h" />
+		<Unit filename="include/Gm1Loader.h" />
 		<Unit filename="include/ImageLoader.h" />
 		<Unit filename="include/JpegLoader.h">
 			<Option compilerVar="CC" />
 		</Unit>
 		<Unit filename="include/PcxLoader.h" />
 		<Unit filename="include/PngLoader.h" />
-		<Unit filename="include/TGXLoader.h" />
+		<Unit filename="include/TGXParser.h" />
 		<Unit filename="include/TgaLoader.h" />
+		<Unit filename="include/TgxLoader.h" />
 		<Unit filename="include/algif/algif.h" />
 		<Unit filename="include/algif/gif.h" />
 		<Unit filename="include/algif/lzw.h" />
 		<Unit filename="include/image_data.h" />
-		<Unit filename="include/init.h" />
 		<Unit filename="include/keyListener.h" />
-		<Unit filename="include/state.h" />
 		<Unit filename="include/view.h" />
 		<Unit filename="res/resources.rc">
 			<Option compilerVar="WINDRES" />
 		</Unit>
 		<Unit filename="src/BmpLoader.cpp" />
-		<Unit filename="src/GM1Loader.cpp" />
+		<Unit filename="src/GM1Parser.cpp" />
 		<Unit filename="src/GifLoader.cpp" />
+		<Unit filename="src/Gm1Loader.cpp" />
 		<Unit filename="src/ImageLoader.cpp" />
 		<Unit filename="src/JpegLoader.cpp" />
 		<Unit filename="src/PcxLoader.cpp" />
 		<Unit filename="src/PngLoader.cpp" />
-		<Unit filename="src/TGXLoader.cpp" />
+		<Unit filename="src/TGXParser.cpp" />
 		<Unit filename="src/TgaLoader.cpp" />
+		<Unit filename="src/TgxLoader.cpp" />
 		<Unit filename="src/algif/algif.c">
 			<Option compilerVar="CC" />
 		</Unit>
@@ -68,10 +70,8 @@
 			<Option compilerVar="CC" />
 		</Unit>
 		<Unit filename="src/image_data.cpp" />
-		<Unit filename="src/init.cpp" />
 		<Unit filename="src/keyListener.cpp" />
 		<Unit filename="src/main.cpp" />
-		<Unit filename="src/state.cpp" />
 		<Unit filename="src/view.cpp" />
 		<Extensions>
 			<code_completion />
