+++ src/main.cpp
@@ -7,19 +7,8 @@
 
 // Libs
 #include <allegro.h>
-
 #include "algif/algif.h"
-#include "state.h"
-#include "init.h"
 #include "view.h"
-
-// Current state object
-state *currentState = NULL;
-
-// Delete game state and free state resources
-void clean_up() {
-  delete currentState;
-}
 
 // Close button handler (enables X button)
 volatile int close_button_pressed = FALSE;
@@ -28,36 +17,6 @@ void close_button_handler(void) {
   close_button_pressed = TRUE;
 }
 END_OF_FUNCTION(close_button_handler)
-
-// Change game state
-void change_state() {
-  //If the state needs to be changed
-  if(nextState != STATE_NULL) {
-    //Delete the current state
-    if(nextState != STATE_EXIT) {
-      delete currentState;
-    }
-
-    //Change the state
-    switch(nextState) {
-      case STATE_INIT:
-        currentState = new init();
-        break;
-      case STATE_VIEW :
-        currentState = new view();
-        break;
-      default:
-        currentState = new init();
-        break;
-    }
-
-    //Change the current state ID
-    stateID = nextState;
-
-    //NULL the next state ID
-    nextState = STATE_NULL;
-  }
-}
 
 // Setup game
 void setup() {
@@ -76,6 +35,12 @@ void setup() {
   // Close button
   LOCK_FUNCTION(close_button_handler);
   set_close_button_callback(close_button_handler);
+
+  // Set graphics mode
+  set_gfx_mode(GFX_AUTODETECT_WINDOWED, 640, 640, 0, 0);
+
+  // Title
+  set_window_title("ALIMG - Allegro Image Viewer");
 }
 
 // Start here
@@ -83,31 +48,18 @@ int main (int argc, char *argv[]) {
   // Setup basic functionality
   setup();
 
-  //Set the current state ID
-  set_next_state(STATE_INIT);
-  change_state();
-  set_next_state(STATE_VIEW);
-  change_state();
-
   // Load images into view
-  view* view_state = dynamic_cast<view*>(currentState);
+  view* view_state = new view();
 
-  // Ensure view_state casted and load images
-  if (view_state != nullptr) {
-    for (int i = 1; i < argc; i++) {
-      view_state -> LoadImage(argv[i]);
-    }
+  for (int i = 1; i < argc; i++) {
+    view_state -> LoadImage(argv[i]);
   }
 
   // Run loop
   while (!key[KEY_ESC] && !close_button_pressed) {
-    currentState -> update();
-    currentState -> draw();
-    change_state();
+    view_state -> update();
+    view_state -> draw();
   }
-
-  //Clean up
-  clean_up();
 
   // End
   return 0;
